lambda/scheduler/requirements.txt
requests==2.25.1
psycopg2-binary==2.9.3
boto3==1.20.46